+++ easier/core/passes/collective_initialization.py
